fix desktop
